feat: unify search components by replacing SearchBox/SearchBoxV2 with SearchInput
Amp-Thread-ID: https://ampcode.com/threads/T-019cd147-a26a-70fc-85e8-4fa61571b0f2
